Tests now check similar audio length rather than byte-by-byte identity
Subtly different versions of libmp3lame can cause different MP3s
to be generated for the same input, causing the tests that check
for the files being identical to fail.  Since we're already
checking that the type of the file (in a file(1) sense) is the
same, it makes sense to do a less strict check, so this commit
checks instead that the length of the audio of the file in
seconds is the same to 2 decimal places.