Compare the length of generated audio files less precisely
With different versions of ffmpeg, there seem to be slight
differences in lengths of the generated files (although
they sound identical).  I haven't investigated this in
detail, but my hypothesis is that this is to do with frame
padding in the MP3, so relax the strictness of the test to
greater than the length of a frame in one of these files.